pyproject.toml
[project]
name = "stn-core"
version = "0.1.0"
description = "STN Core — structural evaluation engine for STN AST"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "stn-lexer",
]

[tool.uv.sources]
stn-lexer = { path = "../STN_lexer", editable = true }

[dependency-groups]
dev = ["pytest>=8"]

[build-system]
requires = ["uv_build>=0.9.21,<0.10.0"]
build-backend = "uv_build"
